created PriceAdjust.java, edit Cinema, TicketPrice and Constants classes
TLDR: PriceAdjust makes things easier to modify shit ltr on

Price Adjust is a interface class to facilitate checking of ticket prices.
Basically anything which interfaces with PriceAdjust will be used to modify ticketprices in the TicketPrice.java file. cause as of now it's vv hard to determine a ticket price without if elses, which isn't good for updating.
PriceAdjust assigned the same class to all the enums in constant that are relavent to determining the price of a ticket